+++ esp-bootloader-esp-idf/build.rs
@@ -24,11 +24,18 @@ fn main() {
             Some(Validator::Enumeration(
                 vec![String::from("8k"), String::from("16k"),String::from("32k"),String::from("64k"),]
             ))
-        ),(
+        ),
+        (
             "esp_idf_version",
             "ESP-IDF version used in the application descriptor. Currently it's not checked by the bootloader.",
             Value::String(String::from("0.0.0")),
             None,
         ),
+        (
+            "partition-table-offset",
+            "The address of partition table (by default 0x8000). Allows you to move the partition table, it gives more space for the bootloader. Note that the bootloader and app will both need to be compiled with the same PARTITION_TABLE_OFFSET value.",
+            Value::Integer(0x8000),
+            Some(Validator::PositiveInteger),
+        )
     ], true);
 }